requirements.txt
nltk
requests
langchain-community
pypdf
langchain-openai
pydantic
fastapi
uvicorn
python-dotenv
langchain-chroma